fix: safetensor models now warn the user they are not supported
The addition of safetensor models in the model reference causes users which use webui.py to configure their worker to be able to select models which are not supported by nataili.

This change will warn the user they are not supported presently and that they will be in the near future.